@@ -11,9 +11,9 @@
     "release:major": "npm version major && npm publish",
     "compile": "rimraf lib/ && babel -d lib/ src/",
     "watch": "babel --watch -d lib/ src/",
-    "jshint": "jshint src/. test/. --config",
-    "mocha": "mocha test/ --compilers js:babel-core/register",
-    "test": "npm run compile && npm run jshint && npm run mocha"
+    "lint": "eslint src/. test/.",
+    "mocha": "mocha test/",
+    "test": "npm run lint && npm run mocha"
   },
   "repository": {
     "type": "git",
@@ -33,26 +33,26 @@
     "url": "https://github.com/thosakwe/feathers-seeder/issues"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">= 10"
   },
   "license": "MIT",
   "directories": {
     "lib": "lib"
   },
   "devDependencies": {
-    "babel-cli": "^6.6.5",
-    "babel-core": "^6.7.2",
-    "babel-plugin-add-module-exports": "^0.1.2",
-    "babel-preset-es2015": "^6.6.0",
-    "debug": "^2.3.3",
-    "feathers-memory": "^2.1.3",
-    "jshint": "^2.9.1",
-    "mocha": "^5.2.0"
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.8.4",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/register": "^7.8.3",
+    "babel-eslint": "^10.0.3",
+    "debug": "^4.1.1",
+    "eslint": "^6.8.0",
+    "feathers-memory": "^4.1.0",
+    "mocha": "^7.0.1"
   },
   "dependencies": {
-    "@feathersjs/cli": "^3.7.3",
-    "@feathersjs/errors": "^3.3.0",
-    "@feathersjs/feathers": "^3.1.7",
+    "@feathersjs/errors": "^4.5.1",
+    "@feathersjs/feathers": "^4.5.1",
     "faker": "^4.1.0",
     "rimraf": "^2.5.2"
   }
