@@ -1,19 +1,15 @@
-'use strict';
+"use strict";
 
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.default = seeder;
+exports["default"] = seeder;
 
-var _errors = require('@feathersjs/errors');
+var _errors = _interopRequireDefault(require("@feathersjs/errors"));
 
-var _errors2 = _interopRequireDefault(_errors);
+var _seeder = _interopRequireDefault(require("./seeder"));
 
-var _seeder = require('./seeder');
-
-var _seeder2 = _interopRequireDefault(_seeder);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
 var debug = require('debug')('feathers-seeder');
 
@@ -24,7 +20,6 @@ function seeder() {
     return function () {
       this.seed = function () {
         debug('Seeder is disabled, not modifying database.');
-
         return Promise.resolve([]);
       };
     };
@@ -36,15 +31,13 @@ function seeder() {
 
   return function () {
     var app = this;
-    var seeder = new _seeder2.default(app, opts);
+    var seeder = new _seeder["default"](app, opts);
 
     app.seed = function () {
-      return seeder.seedApp().then().catch(function (err) {
-        debug('Seeding error: ' + err);
-
-        throw new _errors2.default.GeneralError(err);
+      return seeder.seedApp().then()["catch"](function (err) {
+        debug("Seeding error: ".concat(err));
+        throw new _errors["default"].GeneralError(err);
       });
     };
   };
 }
-module.exports = exports['default'];
